fixed rebase issues
