+++ src/components/posts/form/post-form-section.tsx
@@ -4,12 +4,17 @@ import { Box } from '@material-ui/core'
 import { ActerAvatar } from 'src/components/acter/avatar'
 import { User } from '@schema'
 import { grey } from '@material-ui/core/colors'
-import { PostForm, PostFormProps } from 'src/components/posts/form'
+import {
+  PostForm,
+  PostFormProps,
+  PostFormValues,
+} from 'src/components/posts/form'
 import { Post as PostType } from '@schema'
 import clsx from 'clsx'
 
 export interface PostFormSectionProps extends PostFormProps {
   user: User
+  post?: PostType
 }
 
 export const PostFormSection: FC<PostFormSectionProps> = ({
@@ -22,7 +27,7 @@ export const PostFormSection: FC<PostFormSectionProps> = ({
 
   const handleClick = () => setShowForm(true)
 
-  const handlePostSubmit = (data: PostType) => {
+  const handlePostSubmit = (data: PostFormValues) => {
     setShowForm(false)
     onPostSubmit(data)
   }
